requirements.txt
python-nmap==0.7.1
requests==2.31.0
scapy==2.5.0
cryptography==41.0.1
colorama==0.4.6
tqdm==4.65.0
zxcvbn==4.4.28
